engine: drop e.execIfNotBlocked()
That's where KISS and DRY get into a little conflict. My take is that KISS
wins here:
 * it's five lines of code
 * yeah it's repetitive, but e.execIfNotBlocked() is repetitive as well with
   three lines of code minimum
 * it inherently requires lambdas
 * variables need to be defined and properly managed, some code looks a bit
   strange where we return (smth, error) and smth gets its value in a lambda,
   but error is the result of e.execIfNotBlocked()
 * in most cases it also leads to the need of additional "internal" functions
 * overall we get _less_ code after this patch
 * nothing hidden, clear execution flow

+++ pkg/local_object_storage/engine/container.go
@@ -16,23 +16,17 @@ import (
 //
 // Returns an error if executions are blocked (see BlockExecution).
 func (e *StorageEngine) ContainerSize(cnr cid.ID) (uint64, error) {
-	var (
-		err  error
-		size uint64
-	)
 	if e.metrics != nil {
 		defer elapsed(e.metrics.AddEstimateContainerSizeDuration)()
 	}
 
-	err = e.execIfNotBlocked(func() error {
-		size, err = e.containerSize(cnr)
-		return err
-	})
+	e.blockMtx.RLock()
+	defer e.blockMtx.RUnlock()
 
-	return size, err
-}
+	if e.blockErr != nil {
+		return 0, e.blockErr
+	}
 
-func (e *StorageEngine) containerSize(cnr cid.ID) (uint64, error) {
 	var size uint64
 
 	e.iterateOverUnsortedShards(func(sh hashedShard) (stop bool) {
@@ -58,23 +52,17 @@ func (e *StorageEngine) containerSize(cnr cid.ID) (uint64, error) {
 //
 // Returns an error if executions are blocked (see BlockExecution).
 func (e *StorageEngine) ListContainers() ([]cid.ID, error) {
-	var (
-		res []cid.ID
-		err error
-	)
 	if e.metrics != nil {
 		defer elapsed(e.metrics.AddListContainersDuration)()
 	}
 
-	err = e.execIfNotBlocked(func() error {
-		res, err = e.listContainers()
-		return err
-	})
+	e.blockMtx.RLock()
+	defer e.blockMtx.RUnlock()
 
-	return res, err
-}
+	if e.blockErr != nil {
+		return nil, e.blockErr
+	}
 
-func (e *StorageEngine) listContainers() ([]cid.ID, error) {
 	uniqueIDs := make(map[cid.ID]struct{})
 
 	e.iterateOverUnsortedShards(func(sh hashedShard) (stop bool) {
@@ -103,27 +91,32 @@ func (e *StorageEngine) listContainers() ([]cid.ID, error) {
 
 // DeleteContainer deletes container's objects that engine stores.
 func (e *StorageEngine) DeleteContainer(ctx context.Context, cID cid.ID) error {
-	return e.execIfNotBlocked(func() error {
-		var wg errgroup.Group
+	e.blockMtx.RLock()
+	defer e.blockMtx.RUnlock()
 
-		e.iterateOverUnsortedShards(func(hs hashedShard) bool {
-			wg.Go(func() error {
-				err := hs.Shard.DeleteContainer(ctx, cID)
-				if err != nil {
-					err = fmt.Errorf("container cleanup in %s shard: %w", hs.ID(), err)
-					e.log.Warn("container cleanup", zap.Error(err))
+	if e.blockErr != nil {
+		return e.blockErr
+	}
 
-					return err
-				}
+	var wg errgroup.Group
 
-				return nil
-			})
+	e.iterateOverUnsortedShards(func(hs hashedShard) bool {
+		wg.Go(func() error {
+			err := hs.Shard.DeleteContainer(ctx, cID)
+			if err != nil {
+				err = fmt.Errorf("container cleanup in %s shard: %w", hs.ID(), err)
+				e.log.Warn("container cleanup", zap.Error(err))
 
-			return false
+				return err
+			}
+
+			return nil
 		})
 
-		return wg.Wait()
+		return false
 	})
+
+	return wg.Wait()
 }
 
 func (e *StorageEngine) deleteNotFoundContainers() error {
